% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{get_time_dif}
\alias{get_time_dif}
\title{get_time_dif}
\usage{
get_time_dif(start_time, end_time)
}
\arguments{
\item{start_time}{Start time `Sys.time()`.}

\item{end_time}{End time `Sys.time()`.}
}
\value{
Durantion in HH:MM:SS
}
\description{
get_time_dif
}
